Learning ATTimy85
This repository contains 2 sketches that were instrumental in helping program my first ATTiny85. "Hello Blink" just has an led blink on and off. "Hello Mode" rotates an LED between an off state, an on state, and a blinking state based on the press of a button. The video is just the "Hello Mode" code in action.